+++ src/circuit.rs
@@ -137,9 +137,9 @@ impl VerifierData {
 ///         composer: &mut TurboComposer,
 ///     ) -> Result<(), Error> {
 ///         // Add fixed witness zero
-///         let zero = composer.add_witness_to_circuit_description(BlsScalar::zero());
-///         let a = composer.add_input(self.a);
-///         let b = composer.add_input(self.b);
+///         let zero = composer.zero();
+///         let a = composer.append_witness(self.a);
+///         let b = composer.append_witness(self.b);
 ///         // Make first constraint a + b = c
 ///         composer.poly_gate(
 ///             a,
@@ -168,7 +168,7 @@ impl VerifierData {
 ///             Some(-self.d),
 ///         );
 ///
-///         let e = composer.add_input(self.e);
+///         let e = composer.append_witness(self.e);
 ///         let scalar_mul_result =
 ///             composer.fixed_base_scalar_mul(
 ///                 e, dusk_jubjub::GENERATOR_EXTENDED,
@@ -376,13 +376,13 @@ mod tests {
             &mut self,
             composer: &mut TurboComposer,
         ) -> std::result::Result<(), Error> {
-            let a = composer.add_input(self.a);
-            let b = composer.add_input(self.b);
+            let a = composer.append_witness(self.a);
+            let b = composer.append_witness(self.b);
             // Make first constraint a + b = c
             composer.poly_gate(
                 a,
                 b,
-                composer.allocated_zero(),
+                composer.zero(),
                 BlsScalar::zero(),
                 BlsScalar::one(),
                 BlsScalar::one(),
@@ -397,7 +397,7 @@ mod tests {
             composer.poly_gate(
                 a,
                 b,
-                composer.allocated_zero(),
+                composer.zero(),
                 BlsScalar::one(),
                 BlsScalar::zero(),
                 BlsScalar::zero(),
@@ -406,7 +406,7 @@ mod tests {
                 Some(-self.d),
             );
 
-            let e = composer.add_input(self.e);
+            let e = composer.append_witness(self.e);
             let scalar_mul_result = composer
                 .fixed_base_scalar_mul(e, dusk_jubjub::GENERATOR_EXTENDED);
             // Apply the constrain